+++ VueForm/store/index.js
@@ -66,7 +66,7 @@ export const VueFormStoreParams = {
     isDisabled() {
       const { submitting, validating, dirty } = this.form
 
-      return submitting || validating || !dirty
+      return submitting || validating || (this.formFields.length && !dirty)
     },
 
     allErrors() {